backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery[redis]==5.4.0
redis==5.0.8
httpx==0.27.2
supabase==2.7.4
pydantic==2.8.2
python-dotenv==1.0.1

# LLM (install the one you use)
openai>=1.40.0
anthropic>=0.34.0
